Add files via upload
This is a Gas Station project where the use is asked to choose a pump number, type of fuel, litters and gas station worker ID. The user has and option to start a sale or close at the end of the day.
At the close different information will be printed. 
Quantity sold per pump and type of fuel.
Total litters and $$ per fuel type
Total $$ sold per worker
and mor info.